+++ src/main/java/com/orangeandbronze/leaveapp/repository/LeaveApplicationRepository.java
@@ -6,18 +6,18 @@ import com.orangeandbronze.leaveapp.domain.LeaveApplication;;
 
 public interface LeaveApplicationRepository {
 
-	public LeaveApplication findBy(int leaveId);
+	public LeaveApplication findBy(long leaveId);
 	
 	public List<LeaveApplication> findAllLeaveApplications();
 	
-	public List<LeaveApplication> findLeaveApplicationsByEmployee(int employeeId);
+	public List<LeaveApplication> findLeaveApplicationsByEmployee(long employeeId);
 	
-	public List<LeaveApplication> findLeaveApplicationsForSupervisor(int supervisorId);
+	public List<LeaveApplication> findLeaveApplicationsForSupervisor(long supervisorId);
 	
 	public int updateLeaveStatus(LeaveApplication leaveApplication);
 	
 	public int insert(LeaveApplication leaveApplication);
 	
-	public List<String> findAllCommentsForLeave(int leaveId);
+	public List<String> findAllCommentsForLeave(long leaveId);
 	
 }
